fix: 修复更多 CI 编译警告
修复的问题：
1. file_hash.cpp 缺少 <algorithm> 头文件导致 std::count 未定义
2. file_hash.cpp 中的符号转换警告
   - file.read() 和 file.gcount() 的类型转换
3. event_poll_poll.cpp 中的符号转换警告
   - reserve() 参数的类型转换
4. request_group.cpp 中的类型转换和未使用参数
   - 进度计算的浮点转换
   - engine 参数标记为未使用
5. download_engine_v2.cpp 中的未使用 lambda 捕获和参数
6. http_commands.hpp 中的未使用成员变量
   - 使用 [[maybe_unused]] 属性标记

测试结果：
- 所有 191 个测试通过

+++ packages/libfalcon/src/net/event_poll_poll.cpp
@@ -23,7 +23,7 @@ namespace falcon::net {
 PollEventPoll::PollEventPoll(int max_fds)
     : max_fds_(max_fds)
 {
-    poll_fds_.reserve(max_fds_);
+    poll_fds_.reserve(static_cast<std::size_t>(max_fds_));
     FALCON_LOG_INFO("创建 PollEventPoll: max_fds=" << max_fds_);
 }
